requirements.txt
# FastAPI Server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Image Processing
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0

# PyTorch (CUDA 12.6)
# torch>=2.7.0
# torchvision>=0.18.0
# torchaudio>=2.7.0

# SAM3 - Install from git
# pip install git+https://github.com/facebookresearch/sam3.git

# SAM 3D Objects - Install from git
# pip install git+https://github.com/facebookresearch/sam-3d-objects.git

# Additional dependencies
aiofiles>=23.0.0
httpx>=0.27.0
pydantic>=2.0.0

